Add support for std::chrono::duration. Also split up the test cases into multiple translation units, because the memory and time for compiling are getting a bit ridiculous.
