+++ libraries/psio/tests/test_bool.cpp
@@ -1,0 +1,6 @@
+#include "test_fracpack.hpp"
+
+TEST_CASE("roundtrip bool")
+{
+   test<bool>({false, true});
+}
