+++ libraries/psio/tests/test_char.cpp
@@ -1,0 +1,6 @@
+#include "test_fracpack.hpp"
+
+TEST_CASE("roundtrip char")
+{
+   test<char>({'a', 'b', 'c'});
+}
